+++ keybase/prompter/updateprompt.json
@@ -1,0 +1,6 @@
+{
+	"title": "Keybase Update Version 1.2.3",
+	"message": "The version you are currently running (1.2.2) is outdated.",
+	"description": "Whats New?\nBug fixes\nSecurity fix, see keybase.io",
+	"autoUpdate": true
+}
